Right idea of the implementation, still not working
