+++ config.example.yaml
@@ -83,9 +83,11 @@ chains:
       wormholeChainId: 24
       incentivesAddress: '0x8C8727276725b7Da11fDA6e2646B2d2448E5B3c5'
       bridgeAddress: '0xEe91C335eab126dF5fDB3797EA9d6aD93aeC9722'
-    layerzero:
-      incentivesAddress: '0x495691fafcf2E0c8554bfD6eA67698BCD9795eec'
+    layerZero:
       bridgeAddress: '0x6EDCE65403992e310A62460808c4b910D972f10f'
+      incentivesAddress: '0xDb93559e30F5A3845438DDcf7Ca8A2D6D9005d30'
+      recieverAddress: '0x6EDCE65403992e310A62460808c4b910D972f10f'
+
 
   - chainId: 81457
     name: 'Blast Mainnet'
@@ -96,9 +98,11 @@ chains:
       wormholeChainId: 36
       incentivesAddress: '0x3C5C5436BCa59042cBC835276E51428781366d85'
       bridgeAddress: '0xbebdb6C8ddC678FfA9f8748f85C815C556Dd8ac6'
-    layerzero:
-      incentivesAddress: '0x495691fafcf2E0c8554bfD6eA67698BCD9795eec'
+    layerZero:
       bridgeAddress: '0x6EDCE65403992e310A62460808c4b910D972f10f'
+      incentivesAddress: '0xDb93559e30F5A3845438DDcf7Ca8A2D6D9005d30'
+      recieverAddress: '0x6EDCE65403992e310A62460808c4b910D972f10f'
+
 
   
   - chainId: 8453
@@ -110,7 +114,9 @@ chains:
       wormholeChainId: 30
       incentivesAddress: '0x3C5C5436BCa59042cBC835276E51428781366d85'
       bridgeAddress: '0xbebdb6C8ddC678FfA9f8748f85C815C556Dd8ac6'
-    layerzero:
-      incentivesAddress: '0x495691fafcf2E0c8554bfD6eA67698BCD9795eec'
+    layerZero:
       bridgeAddress: '0x6EDCE65403992e310A62460808c4b910D972f10f'
+      incentivesAddress: '0xDb93559e30F5A3845438DDcf7Ca8A2D6D9005d30'
+      recieverAddress: '0x6EDCE65403992e310A62460808c4b910D972f10f'
+
